Move group components into group folder, minor css changes
