+++ expense-splitter/src/components/expense/CreateExpense.jsx
@@ -1,26 +1,26 @@
 import { nanoid } from 'nanoid'
 import { useForm } from 'react-hook-form'
-import Button from '../Button'
-import {UseDataContext} from '../context/SiteContext'
+import Button from '../ui/Button'
+import { UseDataContext } from '../context/SiteContext'
 
-export default function CreateExpense(){
-    const {
-        handleSubmit,
-        register,
-        formState: { errors },
-      } = useForm()
+export default function CreateExpense() {
+  const {
+    handleSubmit,
+    register,
+    formState: { errors },
+  } = useForm()
 
-    const { expense, handleSetExpense } = UseDataContext()
+  const { expense, handleSetExpense } = UseDataContext()
 
-    const onSubmit = (values) => {
-       handleSetExpense(values)
-      }
+  const onSubmit = (values) => {
+    handleSetExpense(values)
+  }
 
-    console.log(expense)
+  console.log(expense)
 
-    return(
-        <>
-      <h1 className="text-2xl">Create an Expense </h1>
+  return (
+    <div className="mb-5">
+      <h1 className="text-2xl mb-2">Create an Expense </h1>
       <form onSubmit={handleSubmit(onSubmit)}>
         <div className="mb-2">
           <label className="mr-2">Name: </label>
@@ -41,29 +41,29 @@ export default function CreateExpense(){
         </div>
 
         <div className="mb-2">
-            <label htmlFor="category" className="mr-2">
-                Category:
-            </label>
+          <label htmlFor="category" className="mr-2">
+            Category:
+          </label>
 
-            <select 
-                name="category" 
-                {...register('category', {
-                required: "select a category"
-                })}>
-                <option value=""></option>
-                <option value="entertainment">Entertainment</option>
-                <option value="gift">Gift</option>
-                <option value="groceries">Groceries</option>
-                <option value="restaurant">Restaurant</option>
-                <option value="shopping">Shopping</option>
-                <option value="trip">Trip</option>
-                <option value="utilities">Utilities</option>
-                <option value="other">Other</option>
-            </select>
+          <select
+            name="category"
+            {...register('category', {
+              required: 'select a category',
+            })}
+          >
+            <option value=""></option>
+            <option value="entertainment">Entertainment</option>
+            <option value="gift">Gift</option>
+            <option value="groceries">Groceries</option>
+            <option value="restaurant">Restaurant</option>
+            <option value="shopping">Shopping</option>
+            <option value="trip">Trip</option>
+            <option value="utilities">Utilities</option>
+            <option value="other">Other</option>
+          </select>
 
-            {/* {errors.func && <p style={{color:'red'}}> {errors.func.message}</p> } */}
+          {/* {errors.func && <p style={{color:'red'}}> {errors.func.message}</p> } */}
         </div>
-
 
         <div className="mb-2">
           <label className="mr-2">Amount: </label>
@@ -81,27 +81,28 @@ export default function CreateExpense(){
         </div>
 
         <div className="mb-2">
-            <label htmlFor="group" className="mr-2">
-                Group Name:
-            </label>
+          <label htmlFor="group" className="mr-2">
+            Group Name:
+          </label>
 
-            <select 
-                name="group" 
-                {...register('group', {
-                required: "select a group"
-                })}>
-                <option value=""></option>
-                <option value="Beach Lunch">Beach Lunch</option>
-                <option value="gift">Bar Night</option>
-                <option value="groceries">Groceries</option>
-                <option value="restaurant">Restaurant</option>
-                <option value="shopping">Shopping</option>
-                <option value="trip">Trip</option>
-                <option value="utilities">Utilities</option>
-                <option value="other">Other</option>
-            </select>
+          <select
+            name="group"
+            {...register('group', {
+              required: 'select a group',
+            })}
+          >
+            <option value=""></option>
+            <option value="Beach Lunch">Beach Lunch</option>
+            <option value="gift">Bar Night</option>
+            <option value="groceries">Groceries</option>
+            <option value="restaurant">Restaurant</option>
+            <option value="shopping">Shopping</option>
+            <option value="trip">Trip</option>
+            <option value="utilities">Utilities</option>
+            <option value="other">Other</option>
+          </select>
 
-            {/* {errors.func && <p style={{color:'red'}}> {errors.func.message}</p> } */}
+          {/* {errors.func && <p style={{color:'red'}}> {errors.func.message}</p> } */}
         </div>
 
         {/* <div className="mb-2">
@@ -113,9 +114,8 @@ export default function CreateExpense(){
           {errors.description && errors.description.message}
         </div> */}
 
-
         <Button>Submit</Button>
       </form>
-    </>
-    )
+    </div>
+  )
 }
