+++ expense-splitter/src/components/ui/Modal.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Modal() {
+  return (
+    <div className="absolute">
+      <div className="m-auto">MODAL</div>
+    </div>
+  )
+}
